+++ packages/store/abi/MirrorSubscriber.sol/MirrorSubscriber.abi.json.d.ts
@@ -271,6 +271,25 @@ declare const abi: [
     outputs: [];
     stateMutability: "nonpayable";
     type: "function";
+  },
+  {
+    inputs: [
+      {
+        internalType: "bytes4";
+        name: "interfaceId";
+        type: "bytes4";
+      }
+    ];
+    name: "supportsInterface";
+    outputs: [
+      {
+        internalType: "bool";
+        name: "";
+        type: "bool";
+      }
+    ];
+    stateMutability: "pure";
+    type: "function";
   }
 ];
 export default abi;