+++ packages/world/abi/UniqueEntitySystem.sol/UniqueEntitySystem.abi.json.d.ts
@@ -70,6 +70,45 @@ declare const abi: [
   },
   {
     inputs: [];
+    name: "_msgSender";
+    outputs: [
+      {
+        internalType: "address";
+        name: "sender";
+        type: "address";
+      }
+    ];
+    stateMutability: "view";
+    type: "function";
+  },
+  {
+    inputs: [];
+    name: "_msgValue";
+    outputs: [
+      {
+        internalType: "uint256";
+        name: "value";
+        type: "uint256";
+      }
+    ];
+    stateMutability: "pure";
+    type: "function";
+  },
+  {
+    inputs: [];
+    name: "_world";
+    outputs: [
+      {
+        internalType: "address";
+        name: "";
+        type: "address";
+      }
+    ];
+    stateMutability: "view";
+    type: "function";
+  },
+  {
+    inputs: [];
     name: "getUniqueEntity";
     outputs: [
       {
@@ -79,6 +118,25 @@ declare const abi: [
       }
     ];
     stateMutability: "nonpayable";
+    type: "function";
+  },
+  {
+    inputs: [
+      {
+        internalType: "bytes4";
+        name: "interfaceId";
+        type: "bytes4";
+      }
+    ];
+    name: "supportsInterface";
+    outputs: [
+      {
+        internalType: "bool";
+        name: "";
+        type: "bool";
+      }
+    ];
+    stateMutability: "pure";
     type: "function";
   }
 ];