+++ packages/world/abi/WorldRegistrationSystem.sol/WorldRegistrationSystem.abi.json.d.ts
@@ -77,6 +77,38 @@ declare const abi: [
   {
     inputs: [
       {
+        internalType: "uint256";
+        name: "balance";
+        type: "uint256";
+      },
+      {
+        internalType: "uint256";
+        name: "amount";
+        type: "uint256";
+      }
+    ];
+    name: "InsufficientBalance";
+    type: "error";
+  },
+  {
+    inputs: [
+      {
+        internalType: "address";
+        name: "contractAddress";
+        type: "address";
+      },
+      {
+        internalType: "bytes4";
+        name: "interfaceId";
+        type: "bytes4";
+      }
+    ];
+    name: "InterfaceNotSupported";
+    type: "error";
+  },
+  {
+    inputs: [
+      {
         internalType: "string";
         name: "resource";
         type: "string";
@@ -181,6 +213,45 @@ declare const abi: [
     ];
     name: "SystemExists";
     type: "error";
+  },
+  {
+    inputs: [];
+    name: "_msgSender";
+    outputs: [
+      {
+        internalType: "address";
+        name: "sender";
+        type: "address";
+      }
+    ];
+    stateMutability: "view";
+    type: "function";
+  },
+  {
+    inputs: [];
+    name: "_msgValue";
+    outputs: [
+      {
+        internalType: "uint256";
+        name: "value";
+        type: "uint256";
+      }
+    ];
+    stateMutability: "pure";
+    type: "function";
+  },
+  {
+    inputs: [];
+    name: "_world";
+    outputs: [
+      {
+        internalType: "address";
+        name: "";
+        type: "address";
+      }
+    ];
+    stateMutability: "view";
+    type: "function";
   },
   {
     inputs: [
@@ -320,6 +391,25 @@ declare const abi: [
     name: "registerSystemHook";
     outputs: [];
     stateMutability: "nonpayable";
+    type: "function";
+  },
+  {
+    inputs: [
+      {
+        internalType: "bytes4";
+        name: "interfaceId";
+        type: "bytes4";
+      }
+    ];
+    name: "supportsInterface";
+    outputs: [
+      {
+        internalType: "bool";
+        name: "";
+        type: "bool";
+      }
+    ];
+    stateMutability: "pure";
     type: "function";
   },
   {