+++ packages/world/src/interfaces/IWorldErrors.sol
@@ -11,4 +11,6 @@ interface IWorldErrors {
   error FunctionSelectorNotFound(bytes4 functionSelector);
   error ModuleAlreadyInstalled(string module);
   error DelegationNotFound(address delegator, address delegatee);
+  error InsufficientBalance(uint256 balance, uint256 amount);
+  error InterfaceNotSupported(address contractAddress, bytes4 interfaceId);
 }