+++ packages/recs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.0.0-next.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @latticexyz/schema-type@2.0.0-next.8
+  - @latticexyz/utils@2.0.0-next.8
+
 ## 2.0.0-next.7
 
 ### Patch Changes