+++ packages/schema-type/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## 2.0.0-next.8
+
 ## 2.0.0-next.7
 
 ## 2.0.0-next.6